third_party/blink/web_tests/platform/generic/fast/js/mozilla/strict/15.4.4.6-expected.txt

PASS 'use strict'; var a = arr(); [a.pop(), a] threw exception of type TypeError.
PASS var a = arr(); [a.pop(), a] threw exception of type TypeError.
PASS true === true
PASS 'use strict'; var o = obj(); [Array.prototype.pop.call(o), o] threw exception of type TypeError.
PASS var o = obj(); [Array.prototype.pop.call(o), o] threw exception of type TypeError.
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
